Optimize record lookups and use uint32 for `record_offsets` (#38)
* Optimize record lookups in `get_penalty()`

* Use uint32 record offsets and improve overflow checks

* Update pytest graph fixture

+++ cpp/include/seqwin/graph.hpp
@@ -66,7 +66,7 @@ struct Graph {
     /** Sorted by hash. */
     NoInitArray<Edge> edges;
     /** Cumulative global FASTA record offsets by assembly. */
-    std::vector<std::size_t> record_offsets;
+    std::vector<std::uint32_t> record_offsets;
     /** FASTA record IDs of each assembly. */
     std::vector<std::vector<std::string>> ids_by_assembly;
 };
